refactor: improve folder structure
- Moves enums to `app/core/enums.py`.
- Moves models to `app/models/`.
- Change imports
