+++ backend/tests/utils.py
@@ -6,7 +6,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from app.main import app
 from app.data.database import Base
-from app.models import Application, Contact, Interview
+from app.models.models import Application, Contact, Interview
 
 # Use in-memory SQLite for testing
 SQLALCHEMY_DATABASE_URL = "sqlite:///./test.db"
